AoC Day 5 Part 2
This solve took	a while	for me to figure out to solve efficiently.
A brute	force solution worked but took 2 hours of execution.
I am not satisifed with a brute-force solution.

I was able to implement	a way of avoiding many calculations
by minimizing the amount of calculations done on input with
a predictable output.

When starting on the first seed, the next seed to try is increased by N where N=1.
all the way until the last seed in the range.
However, if ALSO the output from seed+N increased by N,
then we can be pretty sure that these fell within the same ranges in all the input output maps.
E.g seed-to-soil -> soil-to-fertilizer -> fertilizer-to-water..

For eveytime this is the case we increase N by the power of 2.
N = [1, 2, 4, 8, ..]
That way we can iterate over the ranges much faster.
Especially considering ranges in the number of millions.

Summary:
We increase the seed with +N for everytime f(seed)+N == f(seed+N).
When this is no longer true we know that somewhere inside the function that seed+N,
fell within another range.
When that happens we set N=1 and start over again until the range is completed..