AoC added command for running AoC inside a container
I had to add some logic to the PuzzleIO class.
For a one line container run command, the project is specified with the dotnet run command.
However, the wayt things are setup is that input and output is read and written
with relative path in mind.

The added logic is a hacky workaround to make it work no matter where you run
the project from, as the execution happens from somewhere between root and the source code.