AoC 2023 Day 23 Part 2 complete (but not fast enough..)
We traverse the paths as before but without the restriction of slopes.
Less restrictions means more permutations (more possible routes to consider).
I wanted to test if the exact solution for part 1 (w/o slopes)  would work at all.
It found the solution in just under 4 hours of runtime :^).
So, I commit this and then I will try to find the correct way to optimize this problem.